route updates

+++ src/place/Http/routes.php
@@ -7,20 +7,13 @@ use Ohio\Spot\Place;
  */
 Route::group([
     'prefix' => 'api/v1',
-    'middleware' => ['api']
+    'middleware' => ['ohio.api']
 ],
     function () {
         Route::get('places/{id}', Place\Http\Controllers\Api\PlacesController::class . '@show');
-        Route::put('places/{id}', Place\Http\Controllers\Api\PlacesController::class . '@update');
-        Route::delete('places/{id}', Place\Http\Controllers\Api\PlacesController::class . '@destroy');
+        Route::put('places/{id}', Place\Http\Controllers\Api\PlacesController::class . '@update')->middleware('ohio.api.admin');
+        Route::delete('places/{id}', Place\Http\Controllers\Api\PlacesController::class . '@destroy')->middleware('ohio.api.admin');
         Route::get('places', Place\Http\Controllers\Api\PlacesController::class . '@index');
-        Route::post('places', Place\Http\Controllers\Api\PlacesController::class . '@store');
-
-        Route::group(['prefix' => 'placegables/{placegable_type}/{placegable_id}'], function () {
-            Route::get('{id}', Place\Http\Controllers\Api\PlacegablesController::class . '@show');
-            Route::delete('{id}', Place\Http\Controllers\Api\PlacegablesController::class . '@destroy');
-            Route::get('', Place\Http\Controllers\Api\PlacegablesController::class . '@index');
-            Route::post('', Place\Http\Controllers\Api\PlacegablesController::class . '@store');
-        });
+        Route::post('places', Place\Http\Controllers\Api\PlacesController::class . '@store')->middleware('ohio.api.admin');
     }
 );
